mv to new file - clean
